% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buildDepends}
\alias{buildDepends}
\title{Return Build-Depends for given packages}
\usage{
buildDepends(regexp = ".")
}
\arguments{
\item{regexp}{A regular expression for the package name(s) with a
default of all (".")}
}
\value{
A character vector containing package names is returned.
}
\description{
The APT Package Management system uses a data-rich caching
structure. This accessor function returns the Build-Depends for
a set of packages matching the given regular expression.
}
\details{
Note that the package lookup uses regular expressions. If only a
single package is desired, append a single \code{$} to terminate
the expression.  \emph{Ie} \code{r-cran-rcpp$} will \emph{not}
return results for \code{r-cran-rcpparmadillo} and
\code{r-cran-rcppeigen}.
}
\examples{
buildDepends("r-cran-rcpp$")
}
\author{
Dirk Eddelbuettel
}
