% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getDepends}
\alias{getDepends}
\title{Return Depends for given packages}
\usage{
getDepends(regexp = ".")
}
\arguments{
\item{regexp}{A regular expression for the package name(s) with a
default of all (".")}
}
\value{
A data frame with four columns listing (source) package, dependend
packages, comparison operator, and, where available, minimal version.
}
\description{
The APT Package Management system uses a data-rich caching
structure. This accessor function returns the Depends for
a set of packages matching the given regular expression.
}
\details{
Note that the package lookup uses regular expressions. If only a
single package is desired, append a single \code{$} to terminate
the expression.  \emph{Ie} \code{r-cran-rcpp$} will \emph{not}
return results for \code{r-cran-rcpparmadillo} and
\code{r-cran-rcppeigen}.
}
\examples{
reverseDepends("r-cran-rcpp$")
}
\author{
Dirk Eddelbuettel
}
