% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getPackages}
\alias{getPackages}
\title{Retrieve Names of All Installable Packages}
\usage{
getPackages(re = ".")
}
\arguments{
\item{Regular}{expression for package name}
}
\value{
A data frame with columns containing the
package name, the installed version (or NA if not installed)
and the section it is installed in (or NA).
}
\description{
The APT Package Management system uses a data-rich caching
structure. This accessor function returns the names of installable
packages for a given regular expression.
}
\author{
Dirk Eddelbuettel
}

