% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hasPackages}
\alias{hasPackages}
\title{Test for Existence of Given Package}
\usage{
hasPackages(pkg)
}
\arguments{
\item{pkg}{Name of the package}
}
\value{
A boolean result is returned
}
\description{
The APT Package Management system uses a data-rich caching '
structure. This accessor function tests whether a given package
exists.
}
\author{
Dirk Eddelbuettel
}

