\name{RcppAPT-package}
\alias{RcppAPT-package}
\alias{RcppAPT}
\docType{package}
\title{
  Rcpp Interface to the APT Package Manager
}
\description{
  Debian-based systems such as Debian, Ubuntu, or their derivatives use
  the APT Package Manager, commonly via programs such as \code{apt-get},
  \code{apt-cache} or other frontends written against the APT libraries.

  This package offers a simple interface from R, mostly so that the
  package management system can be queried.
}
\details{
  The APT library API is described in the package \code{libapt-pkg-doc}.
}
\author{
  Dirk Eddelbuettel

  Maintainer: Dirk Eddelbuettel <edd@debian.org>
}
\references{
  See the \code{libapt-pkg-doc} package on a Debian-based system.
}
\keyword{package}
