% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reverseDepends}
\alias{reverseDepends}
\title{Return Build-Depends for given packages}
\usage{
reverseDepends(regexp = ".")
}
\arguments{
\item{regexp}{A regular expression for the package name(s) with a
default of all (".")}
}
\value{
A character vector containing package names is returned.
}
\description{
The APT Package Management system uses a data-rich caching
structure. This accessor function returns the Build-Depends for
a set of packages matching the given regular expression.
}
\examples{
reverseDepends("r-cran-rcpp$")
}
\author{
Dirk Eddelbuettel
}

