% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{suitable}
\alias{suitable}
\title{Test for Suitability of System}
\usage{
suitable()
}
\value{
A boolean result vector is returned indicating if the 
system is making any sense at all.
}
\description{
The APT Package Management system uses a data-rich caching
structure. This accessor function tests whether a given package
exists.
}
\details{
CRAN does not manage to blacklist this package for builds where it 
has little to no chance of building (macOS amongt them). So we now
build everywhere whether it makes sense or not.
}
\examples{
suitable()
}
\author{
Dirk Eddelbuettel
}
