% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hasPackages}
\alias{hasPackages}
\title{Test for Existence of Given Package}
\usage{
hasPackages(pkg)
}
\arguments{
\item{pkg}{A character vector with name of the package}
}
\value{
A boolean result vector is returned indicating if the 
package at the given position is available.
}
\description{
The APT Package Management system uses a data-rich caching
structure. This accessor function tests whether a given package
exists.
}
\examples{
hasPackages(c("r-base-core", "somethingThatDoesNotExist"))
}
\author{
Dirk Eddelbuettel
}
